requirements.txt
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg==3.2.9